requirements.txt
google-generativeai==0.3.1
termcolor==2.3.0
